requirements.txt
requests
pandas
numpy
scikit-learn
pyyaml
dvclive
matplotlib
dvc
dvc[s3]
